train_models.py
"""
Training script for fashion AI models (No Computer Vision)
"""

import argparse
import yaml
import torch
from torch.utils.data import DataLoader
from models.text_to_fashion import TextToFashionGenerator
from models.style_transfer import StyleTransferModel
from models.trend_predictor import TrendPredictor
from models.fashion_recommender import FashionRecommendationEngine
from utils.data_loader import FashionDataLoader
import logging

def setup_logging():
    logging.basicConfig(
        level=logging.INFO,
        format='%(asctime)s - %(levelname)s - %(message)s',
        handlers=[
            logging.FileHandler('logs/training.log'),
            logging.StreamHandler()
        ]
    )

def load_config(config_path):
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)

def train_text_to_fashion(config, data_loader):
    logging.info("Training Text-to-Fashion model...")
    model = TextToFashionGenerator()
    
    # Get dataloader
    dataset = data_loader.get_fashion_dataset('train')
    dataloader = data_loader.create_dataloader(dataset, config['batch_size'])
    
    # Train model
    model.train(dataloader, config.get('num_epochs', 50))
    logging.info("Text-to-Fashion training completed!")

def train_style_transfer(config, data_loader):
    logging.info("Training Style Transfer model...")
    model = StyleTransferModel()
    
    # Get dataloader for style transfer
    dataset = data_loader.get_style_transfer_dataset('./data/content_images', './data/style_images')
    dataloader = data_loader.create_dataloader(dataset, config['batch_size'])
    
    # Train model
    model.train(dataloader, config.get('num_epochs', 50))
    logging.info("Style Transfer training completed!")

def train_trend_predictor(config, data_loader):
    logging.info("Training Trend Predictor model...")
    model = TrendPredictor()
    
    # Get dataloader
    dataset = data_loader.get_fashion_dataset('train')
    dataloader = data_loader.create_dataloader(dataset, config['batch_size'])
    
    # Train model (simplified without heavy CV)
    logging.info("Trend prediction model initialized with rule-based analysis")
    logging.info("Trend Predictor setup completed!")

def train_fashion_recommender(config, data_loader):
    logging.info("Training Fashion Recommender model...")
    model = FashionRecommendationEngine()
    
    # Get dataloader
    dataset = data_loader.get_fashion_dataset('train')
    dataloader = data_loader.create_dataloader(dataset, config['batch_size'])
    
    # Train collaborative filtering model
    model.train(dataloader, config.get('num_epochs', 50))
    logging.info("Fashion Recommender training completed!")

def main():
    parser = argparse.ArgumentParser(description='Train Fashion AI Models')
    parser.add_argument('--model', type=str, choices=['all', 'text2fashion', 'style_transfer', 'trend_predictor', 'recommender'], 
                        default='all', help='Model to train')
    parser.add_argument('--config', type=str, default='configs/model_config.yaml', help='Config file path')
    parser.add_argument('--data_dir', type=str, default='./data', help='Data directory')
    
    args = parser.parse_args()
    
    setup_logging()
    config = load_config(args.config)
    data_loader = FashionDataLoader(args.data_dir)
    
    if args.model == 'all' or args.model == 'text2fashion':
        train_text_to_fashion(config['text_to_fashion'], data_loader)
    
    if args.model == 'all' or args.model == 'style_transfer':
        train_style_transfer(config['style_transfer'], data_loader)
    
    if args.model == 'all' or args.model == 'trend_predictor':
        train_trend_predictor(config['trend_predictor'], data_loader)
    
    if args.model == 'all' or args.model == 'recommender':
        train_fashion_recommender(config['fashion_recommender'], data_loader)

if __name__ == "__main__":
    main()